Extract skill selector to a standalone component
