+++ pages/join/form.tsx
@@ -1,27 +1,22 @@
 import { NextPage, GetStaticProps } from "next";
 import { array, dict, string } from "typescript-json-decoder";
 import { readFile } from "fs/promises";
+import { SkillMenu } from "components/user-profile/skill-picker";
 import yaml from "js-yaml";
 import OnboardingFormPage, {
-  CompetencyList,
   RegistrationData,
 } from "components/onboarding/form";
 
 type PageProps = {
-  defaultCompetencyList: CompetencyList;
+  skillMenu: SkillMenu;
 };
 
-const Page: NextPage<PageProps> = ({ defaultCompetencyList }) => {
+const Page: NextPage<PageProps> = ({ skillMenu }) => {
   const handleSubmit = async (data: RegistrationData) => {
     console.log(`Submitted data: ${JSON.stringify(data, null, 2)}`);
     await new Promise((r) => setTimeout(r, 4000));
   };
-  return (
-    <OnboardingFormPage
-      competencyList={defaultCompetencyList}
-      onSubmit={handleSubmit}
-    />
-  );
+  return <OnboardingFormPage skillMenu={skillMenu} onSubmit={handleSubmit} />;
 };
 
 async function createUserProfile(data: any): Promise<boolean> {
@@ -38,7 +33,7 @@ async function createUserProfile(data: any): Promise<boolean> {
   }
 }
 
-async function getDefaultCompetencyList(): Promise<Record<string, string[]>> {
+async function getDefaultSkillMenu(): Promise<SkillMenu> {
   const decode = dict(array(string));
   return await readFile("content/competencies.yaml", "utf-8")
     .then((str) => yaml.load(str) as any)
@@ -48,9 +43,9 @@ async function getDefaultCompetencyList(): Promise<Record<string, string[]>> {
 }
 
 export const getStaticProps: GetStaticProps<PageProps> = async () => {
-  const defaultCompetencyList = await getDefaultCompetencyList();
+  const skillMenu = await getDefaultSkillMenu();
   return {
-    props: { defaultCompetencyList },
+    props: { skillMenu },
   };
 };
